+++ changelog.md
@@ -26,6 +26,7 @@
 - fixed forge faction duplication bug where button did not trigger
 - new user profiles failed to save avatar correctly
 - fixed possible error where following the manage guild commands link from /help could 404 if the guild was not registered previously
+- moved forge feedback to the feedback category and out of the forge activity category
 
 ## [1.13.3] - 2026-8-21 Post Spain Updates